mini-dc-digital-twin/pyproject.toml
[project]
name = "dc-digital-twin"
version = "0.1.0"
description = "A small portfolio project inspired by Niagara-style supervisory control for data centers."
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "clickhouse-connect==0.7.19",
    "fastapi==0.115.0",
    "paho-mqtt==2.1.0",
    "python-dotenv==1.0.1",
    "uvicorn==0.30.6",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["app"]
